pubsub-worker/pyproject.toml
[tool.poetry]
name = "pubsub-worker"
version = "0.1.0"
description = ""
authors = ["Jacob Hayes <jacob.r.hayes@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
google-cloud-pubsub = "^1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
